requirements.txt
pyTelegramBotAPI>=4.23.0
python-telegram-bot==21.6
python-binance==1.0.19
pandas==2.0.3
ta==0.11.0
numpy==1.23.5
supabase>=2.6.0
python-telegram-bot[job-queue]>=20.0
websocket-client==1.8.0